langchain-fastapi-chatbot/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
langchain==0.0.284
openai==0.27.8
python-dotenv==1.0.0
pandas==2.2.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
tiktoken==0.4.0
transformers==4.34.0
scikit-learn==1.3.0
pydantic==1.10.12
pytest==7.4.0